Don't show Selenium test menu in productive deployment
Menu and other UI components only required for Selenium test can be
shown or hidden with the new config property 'widgets.selenium'.

Moved some classes to sub-packages, merged two top-level properties
Java classes (deleted WidgetsProperties.java)
