+++ code/widgets/org.eclipse.scout.widgets.client/src/main/java/org/eclipse/scout/widgets/client/ui/forms/BrowserFieldForm.java
@@ -37,8 +37,8 @@ import org.eclipse.scout.rt.platform.text.TEXTS;
 import org.eclipse.scout.rt.platform.util.IOUtility;
 import org.eclipse.scout.rt.platform.util.collection.OrderedCollection;
 import org.eclipse.scout.widgets.client.ResourceBase;
+import org.eclipse.scout.widgets.client.WidgetsClientConfigProperties.ReadOnlyProperty;
 import org.eclipse.scout.widgets.client.WidgetsHelper;
-import org.eclipse.scout.widgets.client.WidgetsProperties.ReadOnlyProperty;
 import org.eclipse.scout.widgets.client.ui.desktop.outlines.IAdvancedExampleForm;
 import org.eclipse.scout.widgets.client.ui.forms.BrowserFieldForm.MainBox.CloseButton;
 import org.eclipse.scout.widgets.client.ui.forms.BrowserFieldForm.MainBox.ExamplesBox;
