+++ code/widgets/org.eclipse.scout.widgets.client/src/main/java/org/eclipse/scout/widgets/client/ui/forms/FileChooserFieldForm.java
@@ -49,8 +49,8 @@ import org.eclipse.scout.rt.platform.text.TEXTS;
 import org.eclipse.scout.rt.platform.util.CollectionUtility;
 import org.eclipse.scout.rt.platform.util.ObjectUtility;
 import org.eclipse.scout.rt.platform.util.collection.OrderedCollection;
+import org.eclipse.scout.widgets.client.WidgetsClientConfigProperties.ReadOnlyProperty;
 import org.eclipse.scout.widgets.client.WidgetsHelper;
-import org.eclipse.scout.widgets.client.WidgetsProperties.ReadOnlyProperty;
 import org.eclipse.scout.widgets.client.ui.desktop.outlines.IAdvancedExampleForm;
 import org.eclipse.scout.widgets.client.ui.forms.FileChooserFieldForm.MainBox.CloseButton;
 import org.eclipse.scout.widgets.client.ui.forms.FileChooserFieldForm.MainBox.GroupBox;
