+++ code/widgets/org.eclipse.scout.widgets.ui.html.app.dev/src/main/resources/config.properties
@@ -25,3 +25,6 @@ scout.auth.credentialsPlaintext=true
 
 # CSP - Demo app uses external images in html field and custom widgets -> allow it
 scout.cspDirective[img-src]=*
+
+# Make UI components required for selenium tests visible
+widgets.selenium=true
