+++ code/widgets/org.eclipse.scout.widgets.ui.html.app.selenium/src/main/resources/config.properties
@@ -30,3 +30,6 @@ scout.client.userArea=${user.home}/org.eclipse.scout.widgets.ui.html.selenium
 
 # CSP: Required for SeleniumJavaScript.injectSeleniumScript()
 scout.cspEnabled=false
+
+# Make UI components required for selenium tests visible
+widgets.selenium=true
